eval.cpp: Milloro la funció display (ara fa un fork).

+++ eval.cpp
@@ -44,24 +44,30 @@ void usage() {
   exit(0);
 }
 
-int display_image(const Image& I) {
+void display(const Image& I) {
   char _templ[] = "evalXXXXXX";
   char *tmpfile = _templ;
   tmpfile = tmpnam(tmpfile);
   outfile = string(tmpfile) + ".pnm";
   I.save_pnm(outfile);
-  int ret;
-  {
-    stringstream sout;
-    sout << "display " << outfile;
-    ret = system(sout.str().c_str());
+
+  // Launch a new process
+  if (fork() == 0) {
+    int ret;
+    // I'm the child
+    {
+      stringstream sout;
+      sout << "display " << outfile;
+      ret = system(sout.str().c_str());
+    }
+
+    // When display finishes, we delete the file
+    {
+      stringstream sout;
+      sout << "rm " << outfile;
+      exit(system(sout.str().c_str()));
+    }
   }
-  {
-    stringstream sout;
-    sout << "rm " << outfile;
-    system(sout.str().c_str());
-  }
-  return ret;
 }
 
 int main(int argc, char *argv[]) {
@@ -74,7 +80,7 @@ int main(int argc, char *argv[]) {
   Image I(width, height, -1.0, 1.0, 1.0, -1.0);
   root->eval(I);  
   if (outfile == "<none>") {
-    return display_image(I);
+    display(I);
   }
   else {
     I.save_pnm(outfile);
